removed older version of formAccessCalendar.vue, added view for debug testViewTeamForm.vue
