+++ src/router/index.js
@@ -21,7 +21,7 @@ import TestSubjectManagement from "@/views/TestSubjectManagement.vue";
 import TeamManagement from "@/views/TeamManagement.vue";
 import PendingTasks from "@/views/TaskCenter/PendingTasks.vue";
 import ShiftManagement from "@/views/shiftManagement.vue";
-import FormAccessCalendar from "@/components/team-form-calendar/formAccessCalendarV2.vue";
+import FormAccessCalendar from "@/components/team-form-calendar/formAccessCalendar.vue";
 
 const routes = [
     {
